requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
Django==4.1.13
django-cors-headers==4.3.1
django-extensions==3.2.3
django-minio-storage==0.5.7
django-storages==1.14.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
minio==7.2.7
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pymongo==3.12.3
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.2.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
